# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Scaffold route, left public so you can check the server is up without a token.",
        "tags": [
          "App"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "`POST /auth/login`: exchanges an email and password for an access token.\n\n`@Body()` hands over the request body as a `LoginDto`; Nest validates it first and\nanswers `400` itself if it is invalid. `@HttpCode(200)` is needed because Nest\nanswers `201 Created` to every POST by default, and a login creates nothing.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "summary": "`GET /auth/me`: returns the user the token belongs to. The frontend calls it when the\npage loads to check that a stored token is still valid and to learn who is logged in.\n\nIt has no `@Public()`, so JwtAuthGuard requires a valid token before this runs.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/applications": {
      "post": {
        "operationId": "ApplicationsController_receive",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The website's application JSON (snake_case fields, values mostly strings). Required: request_id, first_name, last_name, email, phone. Other fields are kept as sent.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationReceivedDto"
                }
              }
            }
          }
        },
        "summary": "`POST /applications`: receives a driver application from the gonzocar.com form.\n\nPublic: applicants have no account. The body is typed `unknown` instead of a DTO class,\nso the global ValidationPipe leaves it alone and ApplicationsService does the (lenient)\nchecks. Answers `201` with the same body for a new or a repeated submission.",
        "tags": [
          "Applications"
        ]
      },
      "get": {
        "operationId": "ApplicationsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Only applications with this status. Omit for all.",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "on_hold"
              ]
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Page number, starting at 1.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Items per page (1 to 100).",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationListDto"
                }
              }
            }
          }
        },
        "summary": "`GET /applications`: one page of applications, newest first.\n\n`@Query()` collects the query string into the DTO and validates it, the same way\n`@Body()` does for a request body. An invalid value (e.g. `status=foo`) is a `400`.",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/{id}": {
      "get": {
        "operationId": "ApplicationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDetailDto"
                }
              }
            }
          }
        },
        "summary": "`GET /applications/:id`: one application with its full details.\n\n`ParseUUIDPipe` rejects an id that is not a UUID with `400` before any query runs\n(Postgres would otherwise fail on it with a 500). A valid but unknown id is a `404`.",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/{id}/approve": {
      "post": {
        "operationId": "ApplicationsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDetailDto"
                }
              }
            }
          }
        },
        "summary": "`POST /applications/:id/approve`: creates the driver and marks the application approved.\n\nPOST because it is an action with side effects, not an edit of a field. `@HttpCode(200)`\nbecause it returns the updated application rather than creating a resource at this URL.",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/{id}/reject": {
      "post": {
        "operationId": "ApplicationsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDetailDto"
                }
              }
            }
          }
        },
        "summary": "`POST /applications/:id/reject`: rejects the application. No driver is created.",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/{id}/hold": {
      "post": {
        "operationId": "ApplicationsController_hold",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDetailDto"
                }
              }
            }
          }
        },
        "summary": "`POST /applications/:id/hold`: puts the application on hold to decide later.",
        "tags": [
          "Applications"
        ]
      }
    },
    "/drivers": {
      "get": {
        "operationId": "DriversController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Only drivers with this status. Omit for all.",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive"
              ]
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Page number, starting at 1.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Items per page (1 to 100).",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverListDto"
                }
              }
            }
          }
        },
        "summary": "`GET /drivers`: one page of drivers, newest first.",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/drivers/{id}": {
      "get": {
        "operationId": "DriversController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverDto"
                }
              }
            }
          }
        },
        "summary": "`GET /drivers/:id`: one driver. `400` for a malformed id, `404` if unknown.",
        "tags": [
          "Drivers"
        ]
      }
    }
  },
  "info": {
    "title": "Car Rental API",
    "description": "Admin and driver API for the car rental platform.",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The account's email address.",
            "maxLength": 254,
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "The account's password, in plain text (it only travels over HTTPS in production)."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "driver"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Signed token the client sends back on every request as `Authorization: Bearer <token>`."
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "ApplicationReceivedDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Always `\"received\"`.",
            "enum": [
              "received"
            ]
          },
          "application_id": {
            "type": "string",
            "description": "Our id for the application. A repeated submission (same `request_id`) gets the same id."
          }
        },
        "required": [
          "status",
          "application_id"
        ]
      },
      "ApplicationListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "on_hold"
            ]
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "description": "As the applicant typed it (usually 10 digits)."
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "When the applicant pressed submit, by their browser's clock."
          },
          "receivedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When our API received it."
          }
        },
        "required": [
          "id",
          "status",
          "firstName",
          "lastName",
          "email",
          "phone",
          "city",
          "submittedAt",
          "receivedAt"
        ]
      },
      "ApplicationListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationListItemDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Number of applications matching the filter, across all pages."
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
      "ApplicationDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "rejected",
              "on_hold"
            ],
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "description": "As the applicant typed it (usually 10 digits)."
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "When the applicant pressed submit, by their browser's clock."
          },
          "receivedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When our API received it."
          },
          "requestId": {
            "type": "string",
            "description": "The website's id for the submission."
          },
          "zip": {
            "type": "string",
            "nullable": true
          },
          "licenseStoragePath": {
            "type": "string",
            "nullable": true,
            "description": "Paths in the client's Supabase `driver-documents` bucket."
          },
          "ratingStoragePath": {
            "type": "string",
            "nullable": true
          },
          "earningsStoragePath": {
            "type": "string",
            "nullable": true
          },
          "payload": {
            "type": "object",
            "additionalProperties": true,
            "description": "The full submission as the website sent it, minus the bot-check token."
          },
          "reviewedById": {
            "type": "string",
            "nullable": true,
            "description": "The admin who last reviewed it. Null while pending."
          },
          "reviewedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "driverId": {
            "type": "string",
            "nullable": true,
            "description": "The driver created on approval. Null unless approved."
          }
        },
        "required": [
          "id",
          "status",
          "firstName",
          "lastName",
          "email",
          "phone",
          "city",
          "submittedAt",
          "receivedAt",
          "requestId",
          "zip",
          "licenseStoragePath",
          "ratingStoragePath",
          "earningsStoragePath",
          "payload",
          "reviewedById",
          "reviewedAt",
          "driverId"
        ]
      },
      "DriverDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "description": "E.164, e.g. `+13125550123`."
          },
          "status": {
            "enum": [
              "active",
              "inactive"
            ],
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the driver was created (the application's approval)."
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "phone",
          "status",
          "createdAt"
        ]
      },
      "DriverListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DriverDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Number of drivers matching the filter, across all pages."
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}